# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSprint 2026 Hackathon API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EventStats {
  participants: number;
  teams: number;
  prizePool: string;
  judges: number;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  priority?: string | null;
  createdAt: string;
}

export type AnnouncementInputPriority = typeof AnnouncementInputPriority[keyof typeof AnnouncementInputPriority];


export const AnnouncementInputPriority = {
  info: 'info',
  warning: 'warning',
  success: 'success',
} as const;

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  priority?: AnnouncementInputPriority;
}

export type AnnouncementUpdatePriority = typeof AnnouncementUpdatePriority[keyof typeof AnnouncementUpdatePriority];


export const AnnouncementUpdatePriority = {
  info: 'info',
  warning: 'warning',
  success: 'success',
} as const;

export interface AnnouncementUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  priority?: AnnouncementUpdatePriority;
}

export interface Registration {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  college: string;
  teamName: string;
  teamSize: number;
  /** @nullable */
  teamMembers?: string | null;
  createdAt: string;
}

export interface RegistrationInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  college: string;
  /** @minLength 1 */
  teamName: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSize: number;
  teamMembers?: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  authenticated: boolean;
}

export interface DashboardStats {
  totalRegistrations: number;
  teamsRegistered: number;
  individualParticipants: number;
  recentRegistrations: Registration[];
}

export type ListRegistrationsParams = {
search?: string;
};