# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeSprint 2026 Hackathon API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Hackathon registrations
  - name: announcements
    description: Announcements
  - name: admin
    description: Admin operations
  - name: stats
    description: Public stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get public event stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"

  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all active announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "401":
          description: Unauthorized

  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List registrations (admin only)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
        "401":
          description: Unauthorized
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Submit a hackathon registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
        "409":
          description: Email already registered

  /registrations/{id}:
    delete:
      operationId: deleteRegistration
      tags: [registrations]
      summary: Delete a registration (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /registrations/export:
    get:
      operationId: exportRegistrations
      tags: [registrations]
      summary: Export registrations as CSV (admin only)
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    EventStats:
      type: object
      required: [participants, teams, prizePool, judges]
      properties:
        participants:
          type: integer
        teams:
          type: integer
        prizePool:
          type: string
        judges:
          type: integer

    Announcement:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        priority:
          type: ["string", "null"]
        createdAt:
          type: string

    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [info, warning, success]

    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [info, warning, success]

    Registration:
      type: object
      required: [id, fullName, email, phone, college, teamName, teamSize, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        college:
          type: string
        teamName:
          type: string
        teamSize:
          type: integer
        teamMembers:
          type: ["string", "null"]
        createdAt:
          type: string

    RegistrationInput:
      type: object
      required: [fullName, email, phone, college, teamName, teamSize]
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
        college:
          type: string
          minLength: 1
        teamName:
          type: string
          minLength: 1
        teamSize:
          type: integer
          minimum: 1
          maximum: 4
        teamMembers:
          type: string

    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AdminSession:
      type: object
      required: [username, authenticated]
      properties:
        username:
          type: string
        authenticated:
          type: boolean

    DashboardStats:
      type: object
      required: [totalRegistrations, teamsRegistered, individualParticipants, recentRegistrations]
      properties:
        totalRegistrations:
          type: integer
        teamsRegistered:
          type: integer
        individualParticipants:
          type: integer
        recentRegistrations:
          type: array
          items:
            $ref: "#/components/schemas/Registration"