# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSprint 2026 Hackathon API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementUpdatePriority } from './announcementUpdatePriority';

export interface AnnouncementUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  priority?: AnnouncementUpdatePriority;
}