# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSprint 2026 Hackathon API
 * OpenAPI spec version: 0.1.0
 */
import type { Registration } from './registration';

export interface DashboardStats {
  totalRegistrations: number;
  teamsRegistered: number;
  individualParticipants: number;
  recentRegistrations: Registration[];
}