# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSprint 2026 Hackathon API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  college: string;
  teamName: string;
  teamSize: number;
  /** @nullable */
  teamMembers?: string | null;
  createdAt: string;
}