# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSprint 2026 Hackathon API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  college: string;
  /** @minLength 1 */
  teamName: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSize: number;
  teamMembers?: string;
}